Refactor playground state handling
